--- lbrynet/core/HashAnnouncer.py
@@ -1,18 +1,0 @@
-class DummyHashAnnouncer(object):
-    def __init__(self, *args):
-        pass
-
-    def run_manage_loop(self):
-        pass
-
-    def stop(self):
-        pass
-
-    def add_supplier(self, *args):
-        pass
-
-    def hash_queue_size(self):
-        return 0
-
-    def immediate_announce(self, *args):
-        pass
