+++ lbrynet/dht/peerfinder.py
@@ -10,7 +10,23 @@ from lbrynet.core.utils import short_hash
 log = logging.getLogger(__name__)
 
 
-class DHTPeerFinder(object):
+class DummyPeerFinder(object):
+    """This class finds peers which have announced to the DHT that they have certain blobs"""
+
+    def run_manage_loop(self):
+        pass
+
+    def stop(self):
+        pass
+
+    def find_peers_for_blob(self, blob_hash):
+        return defer.succeed([])
+
+    def get_most_popular_hashes(self, num_to_return):
+        return []
+
+
+class DHTPeerFinder(DummyPeerFinder):
     """This class finds peers which have announced to the DHT that they have certain blobs"""
     implements(IPeerFinder)
