+++ scripts/encrypt_blob.py
@@ -5,12 +5,11 @@ import sys
 
 from twisted.internet import defer
 from twisted.internet import reactor
-from twisted.protocols import basic
 from twisted.web.client import FileBodyProducer
 
 from lbrynet import conf
 from lbrynet.core import log_support
-from lbrynet.core.HashAnnouncer import DummyHashAnnouncer
+from lbrynet.dht.hashannouncer import DummyHashAnnouncer
 from lbrynet.core.BlobManager import DiskBlobManager
 from lbrynet.cryptstream.CryptStreamCreator import CryptStreamCreator
